+++ lib/srv/desktop/rdp/rdpclient/src/lib.rs
@@ -16,7 +16,6 @@ pub mod cliprdr;
 pub mod errors;
 pub mod piv;
 pub mod rdpdr;
-pub mod scard;
 pub mod util;
 pub mod vchan;
