+++ lib/srv/desktop/rdp/rdpclient/src/rdpdr/scard.rs
